+++ R/zzz.R
@@ -98,6 +98,8 @@ rxode2.api <- names(rxode2::.rxode2ptrs())
   rxode2::.s3register("rxode2::rxUiDeparse", "optimControl")
   rxode2::.s3register("rxode2::rxUiDeparse", "uobyqaControl")
   rxode2::.s3register("rxode2::rxUiDeparse", "tableControl")
+  rxode2::.s3register("rxode2::rxUiDeparse", "agqControl")
+  rxode2::.s3register("rxode2::rxUiDeparse", "laplaceControl")
   .resetCacheIfNeeded()
   .Call(`_rxode2version4`, as.integer(utils::packageVersion("rxode2") >= "4.0.0"))
 }