+++ NEWS.md
@@ -19,6 +19,10 @@
 
 - Tests are optimized to reduce redundant fits and run in parallel.
 
+- `agqControl()` and `laplaceControl()` now have `rxUiDeparse()`
+  methods so they can be saved better in packages like `nlmixr2save`
+  and `shinyMixR`.
+
 # nlmixr2est 5.0.0
 
 - Remove `qs` and change to `qs2`.  This breaks backward